Updated the case sensativity a bit.
My ultimate aim of having the database entries cased and then the code that gets them work out insensibly seems a little unrealistic as it's easier to lowercase the input rather than the keys in the DB. Maybe at a later date when I've looked at the database code.

Shouldn't affect any other modules as it's standalone (and no port issues this time).