Added in a feature to remove underscores ('_') from the ends of nicks.
The aim of this feature is display the same messages for users logged in more that once under the same nick (it's common to see 'nick' and 'nick_', etc. for those with bad connections).

Of course there are users that won't conform to this, but short of messing around with hostnames it's the best I've got so far (perhaps ~qalias nick=alt_nick could be a future addition to this).