add color to up/downvotes; move NSFW to strings
Orange isn't a color widely available on IRC clients, so green has been
substituted.

"NSFW" is fairly transcendal when it comes to languages; Catalan,
French, and German generally do not seem to localise the abbreviation.
However, it has been moved to strings.json on the off chance that a
better translation arises.

This commit has not been tested. Merge at your own risk.