Add new "details" box to random imgur page
This commit adds a new box to the imgur page, separating the
"Loading..." box with a new "details" box, which contains the
information found via the API.

The loading box appears on the left, and disappears when an image is
loaded, to be replaced with a box on the right.

API operations, such as a "report" or "submit to gallery" option, could
appear below the current right-aligned box, in a smaller font.

I'm undecided as to whether this commit is an improvement or not.