Add tracking for a certain hash
This is a fairly simple hack to record any imgur URLs discovered via the
random image facility that match a certain MD5 hash.

A known bug with this commit is that it may crash the bot the first time
the bot tries to add to the database, as the category is undefined. This
can be easily worked around by ensuring that the category is
prepopulated with at least one image before letting the bot add
automatically.

An issue should probably be opened to merge this code into something
reading hashes from a configuration file: it's pretty hardcoded at the
moment.